Adds IGNOREEMPTY=[0|1] to {P,C}BEACON cmd.
When using COMMENTCMD (PBEACON), or INFOCMD (CBEACON), direwolf reports
an error when the command does not produce any stdout, even when the
command exits with status 0.

This commit allows something like:

PBEACON IGNOREEMPTY=1 EVERY=00:30 COMMENTCMD='command_to_print_updates'

Say 'command_to_print_updates' only prints to stdout when its output
would be different from the last time it ran. With the above, direwolf
would run the command as before, but does not print an error when no new
updates are available. An error is still reported if the exit status is
not 0.