decode_aprs.c: Limit __compar_fn_t to Linux
POSIX does not define __compar_fn_t.   Rather than using it, except on
a list systems where it does not exist, invert the conditional so that
it is used when it is known to exist, and on other systems --
including unknown systems -- use a POSIX-compatible invocation.

Probably switching on __linux__ isn't really right, and instead there
should be a feature test in cmake.  Alternatively, since passing
tocall_cmp without a cast doesn't result in warning, because it is the
same type except static, it would be simpler to just drop the use of
__compar_fn_t entirely.