Remove duplicate (likely incorrect) morse definition for '-'
The Morse table contains two conflicting definitions for '-':

1. "-...-" on L101 (henceforth "3dit")
2. "-....-" on L108 ("4dit")

The order of the list meant that 3dit was used as it came first.
However it also used 3dit for `=` (L107).

Multiple sources, most notably the International Telecommunication Union
Recommendation for Morse code (ITU-R;
https://www.itu.int/dms_pubrec/itu-r/rec/m/R-REC-M.1677-1-200910-I!!PDF-E.pdf),
use 4dit for `-`, and I found nobody else using 3dit.

As such, especially given the conflicting definitions, I think 3dit is
an error, and this commit removes it.